aider: Change handleDuplicateTestCase to do a deep copy.
# Aider chat conversation:

USER: Change handleDuplicateTestCase to do a deep copy

ASSISTANT: To perform a deep copy of the `testCase` object in the `handleDuplicateTestCase` function, we need to replace the spread operator (`...`) with `JSON.parse(JSON.stringify(testCases[index]))`. This will create a new object with the same values, but it won't share references with the original object.

Here is the change:

```tsx
src/web/webui/pages/setup/TestCasesSection.tsx
<<<<<<< ORIGINAL
  const handleDuplicateTestCase = (event: React.MouseEvent, index: number) => {
    event.stopPropagation();
    const duplicatedTestCase = { ...testCases[index] };
    setTestCases([...testCases, duplicatedTestCase]);
  };
=======
  const handleDuplicateTestCase = (event: React.MouseEvent, index: number) => {
    event.stopPropagation();
    const duplicatedTestCase = JSON.parse(JSON.stringify(testCases[index]));
    setTestCases([...testCases, duplicatedTestCase]);
  };
>>>>>>> UPDATED
```

This change will ensure that the duplicated test case is a deep copy of the original, so changes to the copy won't affect the original.