aider: Added error feedback to the user using `alert()` in the catch block.
# Aider chat conversation:

USER: Provide some feedback to the user on error.  alert() is fine

ASSISTANT: To provide feedback to the user when an error occurs, we can use the `alert()` function to display an error message. We will modify the `runTestSuite` function in the `RunTestSuiteButton.tsx` file to call `alert()` with the error message when an error is caught.

Here are the changes:

```tsx
src/web/nextui/src/app/setup/RunTestSuiteButton.tsx
<<<<<<< HEAD
    } catch (error) {
      console.error(error);
      setIsRunning(false);
    }
=======
    } catch (error) {
      console.error(error);
      setIsRunning(false);
      alert(`An error occurred: ${error.message}`);
    }
>>>>>>> updated
```

This change adds a call to `alert()` with the error message in the catch block. This will display an alert dialog with the error message when an error is caught.