aider: Remove onClick handler from user.email MenuItem.
# Aider chat conversation:

USER: make the user.email menuitem not clicky

ASSISTANT: To make the user email MenuItem not clickable, we can remove the onClick handler from it.

Here is the change:

```tsx
src/web/nextui/src/app/components/LoggedInAs.tsx
<<<<<<< HEAD
        <MenuItem>{user.email}</MenuItem>
=======
        <MenuItem disabled>{user.email}</MenuItem>
>>>>>>> updated
```

In the updated code, I've added the `disabled` prop to the MenuItem. This will make the MenuItem appear in a disabled state, which means it can't be clicked or focused.