Trim the end of the model-graded-closedqa response
Fixes #1308

When gemma2:9b is used for model-graded-closedqa, the response often
ends with a white space and a newline. It results in a malformed
response error. In general, this problem can occur with models other
than gemma2, but seems to occur more often with gemma2. The current
implementation of response check is too strict.

This patch trim the end of the model-graded-closedqa response and relax
the check.