feat: enable skipping nunjucks rendering
Ideally prompts are tested with the actual files that are used in
production, for that reason I render the prompt in my prompt function
using my own templating tools

Originally I just let it flow through nunjucks afterwards because it
didn't break anything, but my prompt embeds user code and
that causes nunjucks to explode. This change enables users to
easily skip the rendering process if required.