feat: persist search in url
While the search state is not lost during navigation, if i want to for
example link the page to someone else with the correct search state, or
get a llm to extract data from the page.

Being able to modify the search in the url solves both problems.

right now evalId is being used in a query paramter format, but the page also
supports evalId in the url fragment approach.

when updating the search query paramter we include the previous qp with

`const searchEntries = Object.fromEntries(searchParams.entries());`

when navigating we keep the search qp with

`search: window.location.search,`, i changed the navigate function to use
the url fragment versus the qp as it like a nicer url to me, also prevents
having to do a bunch of work to merge the search qp and the evalId qp