fix: incorrect evalId socket mode /evals
eval data is fetched seperately then the latest evalId,

evalId is fetch using /results and evalData uses the socket init.
/results is not sorted by created at, this means that the ui can have a
random evalId based on sqllite default ordering while the data is correct.

this commits update the source of truth to be the same for the data
and the eval by returning the evalId along side the data.

but the latest eval for non socket instances will also be wrong since
it relies on the same ordering

eval grid is unaffected as the table does local ordering but would break
if it was server paginated